Refactor: move error handler into errors module
